% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dodge-package.R
\docType{package}
\name{Dodge-package}
\alias{Dodge-package}
\alias{Dodge}
\title{Acceptance sampling functions}
\description{
A number of sampling plans can be compared for their operating
characteristics and other commonly used functions.
}
\details{
\tabular{ll}{ Package: \tab Dodge\cr Type: \tab Package\cr Version: \tab
0.8\cr Date: \tab 2013-09-18\cr License: \tab GPL\cr LazyLoad: \tab yes\cr }
}
\references{
Dodge
}
\author{
Raj Govindaraju and Jonathan Godfrey

Maintainer: Jonathan Godfrey <a.j.godfrey@massey.ac.nz>
}
\keyword{package}
