% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acceptance.R
\name{plot.AccSampPlan}
\alias{plot.AccSampPlan}
\alias{plot.CurtSampPlan}
\alias{plot.SeqSampPlan}
\title{plot methods for the Dodge package}
\usage{
\method{plot}{AccSampPlan}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class AccSampPlan, CurtSampPlan, or SeqSampPlan}

\item{y}{ignored}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Creates plots for analysing the design of an acceptance sampling procedure.
}
\details{
At this stage the \code{plot.AccSampPlan} method only plots the Operating
Characteristic (OC) curve, the Average (AOQ) and (ATI) against the
proportion (p) of product that is nonconforming. It also plots the curtailed
sample size or the average sample number (ASN) against p. Further
development is still required.
}
\examples{

Plan1 = SSPlanBinomial(1000, 20,1, Plots=FALSE)
plot(Plan1)

}
\author{
Jonathan Godfrey with some assistance from Raj Govindaraju
}
