% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acceptance.R
\name{print.AccSampPlan}
\alias{print.AccSampPlan}
\alias{print.CurtSampPlan}
\alias{print.SeqSampPlan}
\title{print methods for the Dodge package}
\usage{
\method{print}{AccSampPlan}(x, ...)
}
\arguments{
\item{x}{an object of class AccSampPlan, CurtSampPlan, or SeqSampPlan}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Adds to the base functionality for the print() command. The accompanying
\code{plot} methods are more sophisticated.
}
\details{
These methods print the most necessary elements of the corresponding
objects.
}
\seealso{
The corresponding plot method is far more interesting. See
\code{\link{plot.AccSampPlan}} for example.
}
\author{
Jonathan Godfrey
}
