% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequential.R
\name{SeqDesignBinomial}
\alias{SeqDesignBinomial}
\alias{SeqDesign}
\title{Create a sequential sampling plan}
\usage{
SeqDesignBinomial(N = NULL, AQL, alpha, LQL, beta, Plots = TRUE)
}
\arguments{
\item{N}{the lot size, ignored for the design of the plan unless the
underlying distribution is hypergeometric}

\item{AQL}{Acceptable quality level}

\item{alpha}{producer's risk}

\item{LQL}{Limiting quality level}

\item{beta}{consumers' risk}

\item{Plots}{logical stating if the sequential chart should be plotted}
}
\description{
Selects the appropriate sequential sampling plan from the given inputs. The
only distribution that has been used in functions thus far is the binomial,
but further development is expected.
}
\author{
Raj Govindaraju and Jonathan Godfrey
}
