% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequential.R
\name{SequentialBinomial}
\alias{SequentialBinomial}
\alias{Sequential}
\title{Attribute Sequential Sampling Plans}
\usage{
SequentialBinomial(x, Plots = TRUE)
}
\arguments{
\item{x}{an object of class SeqSampPlan, or at least having the same
elements as one.}

\item{Plots}{logical indicating if the four plots should be returned}
}
\description{
Designs an attribute sequential sampling plan for given AQL, alpha, LQL, and
beta. The user can request plots describing the performance of the plan.
}
\examples{

PlanDesign=SeqDesignBinomial(AQL=0.01, alpha=0.05, LQL=0.04, beta=0.05, Plots=FALSE)
SequentialBinomial(PlanDesign)

}
\author{
Raj Govindaraju with minor editing by Jonathan Godfrey
}
