% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single.R
\name{SSPDesignBinomial}
\alias{SSPDesignBinomial}
\alias{SSPDesign}
\alias{SSPDesignPoisson}
\title{Single Sampling Plan Designs}
\usage{
SSPDesignBinomial(AQL, alpha, LQL, beta)
}
\arguments{
\item{AQL}{Acceptable quality level}

\item{alpha}{producer's risk}

\item{LQL}{Limiting quality level}

\item{beta}{consumers' risk}
}
\description{
Design a single sampling plan for given AQL, alpha, LQL, and beta. Currently
there are functions for the binomial and Poisson distributions.
}
\examples{

SSPDesignBinomial(0.01, 0.05, 0.04, 0.05)
SSPDesignPoisson(0.01, 0.05, 0.04, 0.05)

}
\references{
Dodge, H.F. and Romig, H.G. (1959) \dQuote{Sampling Inspection
Tables, Single and Double Sampling}, Second edition, John Wiley and Sons, New
York.
}
\author{
Raj Govindaraju with minor editing by Jonathan Godfrey
}
