% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSP.R
\name{VSPDesign}
\alias{VSPDesign}
\title{Variable Sampling Plan Design}
\usage{
VSPDesign(AQL, alpha, LQL, beta)
}
\arguments{
\item{AQL}{Acceptable quality level}

\item{alpha}{producer's risk}

\item{LQL}{Limiting quality level}

\item{beta}{consumers' risk}
}
\description{
Design the variable sampling plan for given AQL, alpha, LQL, and beta.
}
\examples{

VSPDesign(AQL=0.01, alpha=0.05, LQL=0.04, beta=0.05)

}
\author{
Raj Govindaraju with minor editing by Jonathan Godfrey
}
\keyword{~kwd1}
\keyword{~kwd2}
