% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSP.R
\name{VSPKnown}
\alias{VSPKnown}
\alias{VSPUnknown}
\title{Variable Sampling Plans}
\usage{
VSPKnown(N, n, k, Pa = seq(0, 1, 0.001), Plots = TRUE)
}
\arguments{
\item{N}{the lot size}

\item{n}{the sample size}

\item{k}{the acceptability constant}

\item{Pa}{fraction nonconforming}

\item{Plots}{logical indicating whether the four plots are required}
}
\description{
Variable sampling plans for known and unknown sigma, evaluated for given
parameters.
}
\examples{

VSPKnown(1000, 20,1)
VSPUnknown(1000, 20,1)

}
\author{
Raj Govindaraju with minor editing by Jonathan Godfrey
}
\keyword{~kwd1}
\keyword{~kwd2}
